Implements the following: - Homepage with introduction, menu and list of configuration settings - Business logic to add and query reservations made by ID and date - Unit tests for the business logic code - View to list all reservations - View to add new reservations - Saves the reservations to a in-memory object in the Application object

+++ Reservations/ReservationsCommons/IConfigReader.cs
@@ -13,7 +13,7 @@ namespace ReservationsCommons {
         /// means we are accepting overbooking. So, a value of 1.2 means we are accepting bookings for 
         /// 20% more seats than we actually have.
         /// </summary>
-        float BookingTop { get; }
+        decimal BookingTop { get; }
 
     }
 }
